requirements.txt
streamlit
fitz
openai
sqlite3
faiss-cpu
sentence-transformers
sklearn
